cv-evaluation-engine/cv-evaluation-engine/requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7
Jinja2==3.1.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.7.2
joblib==1.3.2
gunicorn==21.2.0
